sll changes

+++ src/proto-snmp.c
@@ -471,10 +471,10 @@ snmp_set_cookie(unsigned char *px, size_t length, uint64_t seqno)
     return 0;
 }
 
-#define TWO_BYTE       ((~0)<<7)
-#define THREE_BYTE     ((~0)<<14)
-#define FOUR_BYTE      ((~0)<<21)
-#define FIVE_BYTE      ((~0)<<28)
+#define TWO_BYTE       ((unsigned long long)(~0)<<7)
+#define THREE_BYTE     ((unsigned long long)(~0)<<14)
+#define FOUR_BYTE      ((unsigned long long)(~0)<<21)
+#define FIVE_BYTE      ((unsigned long long)(~0)<<28)
 
 
 /****************************************************************************
